requirement.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
cffi==1.14.4
colorama==0.4.3
distlib==0.3.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
pycparser==2.20
pylint==2.5.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
virtualenv==20.0.26
wrapt==1.12.1
